instrument_classifier_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
librosa==0.10.1
tensorflow==2.13.0
numpy==1.24.3
scipy==1.11.4
pydub==0.25.1
python-multipart==0.0.6
Pillow==10.1.0
django-cors-headers==4.3.1
